wip convert genesis schema
